+++ controllers/products/addContact.js
@@ -1,19 +1,7 @@
 const contactsOperation = require('../../models/contacts');
-const Joi = require('joi');
-
-const contactsSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
 
 const addContact = async (req, res, next) => {
   try {
-    const { error } = contactsSchema.validate(req.body);
-    if (error) {
-      error.status = 400;
-      throw error;
-    }
     const newContact = await contactsOperation.addContact(req.body);
     console.log(newContact);
     res.status(201).json({
