+++ controllers/products/updateContact.js
@@ -1,19 +1,7 @@
 const contactsOperation = require('../../models/contacts');
-const Joi = require('joi');
-
-const contactsSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
 
 const updateContact = async (req, res, next) => {
   try {
-    const { error } = contactsSchema.validate(req.body);
-    if (error) {
-      error.status = 400;
-      throw error;
-    }
     const { contactId } = req.params;
     const updateContact = await contactsOperation.updateContact(
       contactId,
