+++ routes/api/contacts.js
@@ -1,5 +1,9 @@
 const express = require('express');
 const { contactsOperation: ctrl } = require('../../controllers');
+const { validation } = require('../../middlewares');
+const { contactsSchema } = require('../../schema');
+
+const validateMiddleware = validation(contactsSchema);
 
 const router = express.Router();
 
@@ -7,10 +11,10 @@ router.get('/', ctrl.getAll);
 
 router.get('/:contactId', ctrl.getById);
 
-router.post('/', ctrl.addContact);
+router.post('/', validateMiddleware, ctrl.addContact);
 
 router.delete('/:contactId', ctrl.removeContact);
 
-router.put('/:contactId', ctrl.updateContact);
+router.put('/:contactId', validateMiddleware, ctrl.updateContact);
 
 module.exports = router;
