+++ schema/index.js
@@ -1,0 +1,3 @@
+const contactsSchema = require('./contacts');
+
+module.exports = { contactsSchema };
